chore(tests): consolidate fake_lancedb fixture into conftest.py
Moves the _CapturingBuilder / _FakeTable / _FakeDB / _CapturingVoyage
helpers and the fake_lancedb fixture from test_search_since.py into
tests/conftest.py so pytest auto-discovers the fixture everywhere.

Fixes the collection-time ModuleNotFoundError in test_search_expansion.py
(previously imported via `from tests.test_search_since import fake_lancedb`,
which fails because tests/ is not a Python package). test_search_expansion.py
now uses the fixture by name only, no cross-test import needed.

Net change: -2 lines. No test behavior change. Flagged in PR #33 as a
follow-up; this is the follow-up.
